% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateDataframe.R
\name{CreateDataframe}
\alias{CreateDataframe}
\title{Create a data frame}
\usage{
CreateDataframe(mat, row.data, col.data, nLVs)
}
\arguments{
\item{mat}{A matrix}

\item{row.data}{Data}

\item{col.data}{Data}

\item{nLVs}{Number of latent variables}
}
\value{
A data frame
}
\description{
Create a data frame
}
\details{
Create a data frame
}
\examples{
CreateDataframe(mat=matrix(1:6, ncol=2), row.data=data.frame(r=1:3),
col.data=data.frame(c=1:2), nLVs=1)
}
