% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HierOrd-package.R
\docType{package}
\name{HierOrd-package}
\alias{HierOrd}
\alias{HierOrd-package}
\title{HierOrd: Hierarchical Ordination, using INLA}
\description{
This does not yet work. But some of the functions might.
}
\keyword{internal}
