% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeFormula.R
\name{MakeFormula}
\alias{MakeFormula}
\title{make formulae for row & column effects}
\usage{
MakeFormula(
  nlv,
  linnames = NULL,
  factnames = NULL,
  factprior =
    "list(theta = list(prec = \\"loggamma\\", param = c(1, 1), initial = 1, fixed = TRUE)",
  epsprior = "list(prec = list(prior = \\"pc.prec\\", param = c(1, 0.01)))",
  epsname = ""
)
}
\arguments{
\item{nlv}{Number of latent variables}

\item{linnames}{names of variables that will be linear, defaults to NULL,}

\item{factnames}{names of variables that will be factors, defaults to NULL}

\item{factprior}{prior for hyperparameters for precision of factors. Defaults to "list(theta = list(prec = \"loggamma\", param = c(1, 1), initial = 1, fixed = TRUE)",}

\item{epsprior}{prior for hyperparameters for residual precision. Defaults to "list(prec = list(prior = \"pc.prec\", param = c(1, 0.01)))",}

\item{epsname}{Name of residual error}
}
\value{
A formula
}
\description{
make formulae for row & column effects
}
\details{
Writes a formula for row or column effects. Adds weight terms etc.
}
\examples{
MakeFormula(nlv=1, linnames=NULL, factnames=NULL, epsname="")
}
