% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeLincombs.R
\name{MakeLincomb}
\alias{MakeLincomb}
\title{make a linear combination for LVs}
\usage{
MakeLincomb(covs, lvname = "LV1", lcname = "lc", epsname = "eps", N = NULL)
}
\arguments{
\item{covs}{covariates}

\item{lvname}{Latent variable name. Defaults to "LV1",}

\item{lcname}{Linear combination name. Defaults to "lc"}

\item{epsname}{Residual error name. Defaults to "eps",}

\item{N}{Nu,nber of rows. Defaulta to NULL}
}
\value{
Linear combination from INLA
}
\description{
make a linear combination for LVs
}
\details{
make linear combinations for LVs
}
\examples{
\dontrun{
MakeLincomb(covs=NULL, lvname="LV1", lcname="lc", epsname="eps", N=3)
}
}
