% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeLincombs.R
\name{MakeLincombs}
\alias{MakeLincombs}
\title{make linear combinations for LVs}
\usage{
MakeLincombs(covar.df, nlv = 2, EpsName = "eps", N = NULL)
}
\arguments{
\item{covar.df}{Data frame of covaraites. Can be NULL, but N must be specified}

\item{nlv}{Number of latent variables. Defaults to 2}

\item{EpsName}{Name for residual error. Defaults to "eps"}

\item{N}{Number of rows, if covar.df is NULL}
}
\value{
Linear combinations from INLA
}
\description{
make linear combinations for LVs
}
\details{
make linear combinations for LVs
}
\examples{
MakeLincombs(covar.df=NULL, nlv=2, EpsName="eps", N=3)
}
