% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotateLVs.R
\name{rotateLVs}
\alias{rotateLVs}
\title{rotate a latent variable}
\usage{
rotateLVs(mat, rotation.fn = "svd", scale = FALSE)
}
\arguments{
\item{mat}{matrix to rotate}

\item{rotation.fn}{Function in GPArotation to calculate rotation, or svd. Defaults to svd.}

\item{scale}{Should rotated matrix be scaled? Defaults to FALSE}
}
\value{
Rotated latenr variables
}
\description{
rotate a latent variable
}
\details{
Function to rotate a latent variable, defaults to Varimax. Methods in GPArotation package
}
\examples{
\dontrun{
rotateLVs <- function(matrix(1:6, ncol=2), rotation.fn="svd", scale=FALSE)
}
}
