% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMCMC.R
\name{RunMCMC}
\alias{RunMCMC}
\title{Function to run one chain of MCMC}
\usage{
RunMCMC(
  data,
  row.form,
  col.form,
  rowLinComb,
  colLinComb,
  nBurnin,
  nIter,
  family,
  inlaobj = FALSE
)
}
\arguments{
\item{data}{data (full object)}

\item{row.form}{row formula}

\item{col.form}{column formula}

\item{rowLinComb}{linear combinationsd for rows}

\item{colLinComb}{linear combinationsd for columns}

\item{nBurnin}{number of iterations as burn in}

\item{nIter}{number of iterations to keep, after burn in}

\item{family}{family, passed to inla()}

\item{inlaobj}{Should the INLA objects be returned? Defaults to FALSE}
}
\value{
A list. Details to follow
}
\description{
Function to run one chain of MCMC
}
\details{
This runs the MCMC for one chain: it should be an internal function, I guess.
}
\examples{
\dontrun{
UpdateScores(scores, data, rowcol, names)
}

}
