% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimLinComb.R
\name{SimLinComb}
\alias{SimLinComb}
\title{simulate linear combinations.}
\usage{
SimLinComb(mod, n = 1)
}
\arguments{
\item{mod}{INLA Model to simulate from}

\item{n}{Number of simulations, defaluts to 1.}
}
\value{
Matrix of LVs
}
\description{
simulate linear combinations.
}
\details{
Simulates LVs
}
\examples{
\dontrun{
SimLinComb(mod, n=1)
}
}
