% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateData.R
\name{SimTrueScores}
\alias{SimTrueScores}
\title{Simulate true row/column scores}
\usage{
SimTrueScores(eff, covs, scale = FALSE)
}
\arguments{
\item{eff}{Proportion of each LV explained by covariates (I think!). Length equals number of LVs}

\item{covs}{Covaraites}

\item{scale}{Should the results be scale to unit variance? Defaults to FALSE}
}
\value{
A matrix of scores, with some rows and columns. This is used in SimulateData()
If scale=TRUE, the standard deviation before scaling is set to be the attribute StdDev
}
\description{
Simulate true row/column scores
}
\details{
This simulates row or column scores.
}
\examples{
\dontrun{
SimTrueScores(eff, covs, scale=FALSE)
}
}
