% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateScores.R
\name{UpdateScores}
\alias{UpdateScores}
\title{Function to update latent variables}
\usage{
UpdateScores(scores, data, rowcol, names)
}
\arguments{
\item{scores}{Scores}

\item{data}{Data}

\item{rowcol}{Row or column? Must be either "Row" or "Col"}

\item{names}{Names}
}
\value{
An updated data frame
}
\description{
Function to update latent variables
}
\details{
update latent variables, i.e. import new LV and then update weights
}
\examples{
\dontrun{
UpdateScores(scores, data, rowcol, names)
}

}
