% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitHierOrd.R
\name{FitHierOrd}
\alias{FitHierOrd}
\title{Function to fit a hierarchical ordination}
\usage{
FitHierOrd(
  Y,
  rowdata,
  coldata,
  nLVs = 1,
  family,
  RowCovHyper = "list(prec=list(prior = \\"pc.prec\\", param = c(1, 0.0001)))",
  EpsHyper = "list(prec=list(prior = \\"pc.prec\\", param = c(1, 0.0001)))",
  NBurnin = 20,
  NIter = 100,
  NChains = 1,
  NCores = NULL,
  parallel = TRUE,
  INLAobj = FALSE
)
}
\arguments{
\item{Y}{the matrix of data}

\item{rowdata}{row covariates}

\item{coldata}{column covariates}

\item{nLVs}{number of latent variables, defaults to 1,}

\item{family}{family}

\item{EpsHyper}{distributon of hyperparameters for residual variation, Defaults to "list(prec=list(prior = \"pc.prec\", param = c(1, 0.0001)))",}

\item{NBurnin}{Number of iterations as burnin, default 20}

\item{NIter}{Number of iterations after burnin, default 100}

\item{NChains}{Number of chains to run, default 1}

\item{NCores}{Number of cores to use if more than one chain is run. If NULL (the default),
set to equal NChains.}

\item{parallel}{Should the MCMC be paralellised if there is more than 1 chain? Defaults to TRUE}

\item{INLAobj}{Logical, should the INLA objects be kept? Default FALSE}
}
\value{
a list (more later...)
}
\description{
Function to fit a hierarchical ordination
}
\details{
Fit.a hierarchical model

Parallelisation is done one chain per cluster.
}
\examples{
\dontrun{
UpdateScores(scores, data, rowcol, names)
}

}
