% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulateData.R
\name{SimulateData}
\alias{SimulateData}
\title{Function to simulate data from a hierarchical ordination model, assuming a Poisson distribution}
\usage{
SimulateData(NRows = 10, NCols = 5, NLVs = 1, ColEff, RowEff, Sigma, Intercept)
}
\arguments{
\item{NRows}{Number of rows, defaults to 10}

\item{NCols}{Number of rows, defaults to 5}

\item{NLVs}{Number of latent variables, defaults to 1}

\item{ColEff}{Vector of Column effects}

\item{RowEff}{Variance of row effects}

\item{Sigma}{Scale for latent variables: standard deviation}

\item{Intercept1}{Intercept}
}
\value{
A list with elements Counts (response), RowCov (row covariates), and ColCov (column covariates)
}
\description{
Function to simulate data from a hierarchical ordination model, assuming a Poisson distribution
}
\details{
This simulates data from a hierarchical ordination, assuming a Poisson distributed response.
}
\examples{
SimulateData(NRows = 10, NCols = 5, NLVs = 1,
                         ColEff=0, RowEff=0 ,Sigma=0.1, Intercept=1)
}
